requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
python-dotenv==1.0.0
boto3==1.34.14
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
mangum==0.17.0